+++ pages/_document.tsx
@@ -2,8 +2,6 @@ import React from 'react'
 import Document, { Html, Head, Main, NextScript } from 'next/document'
 import Script from 'next/script'
 
-import config from '../src/config'
-
 export default class MyDocument extends Document {
   render() {
     return (
@@ -250,24 +248,7 @@ export default class MyDocument extends Document {
             href="/splash/apple-splash-1136-640.jpg"
             media="(device-width: 320px) and (device-height: 568px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)"
           />
-          <Script
-            async
-            src={`https://www.googletagmanager.com/gtag/js?id=${config.googleAnalyticsId}`}
-          />
-          <Script
-            defer
-            dangerouslySetInnerHTML={{
-              __html: `
-                    window.dataLayer = window.dataLayer || [];
-                    function gtag(){dataLayer.push(arguments);}
-                    gtag('js', new Date());
 
-                    gtag('config', '${config.googleAnalyticsId}', {
-                      page_path: window.location.pathname,
-                    });
-                  `,
-            }}
-          />
           <link rel="alternate" type="application/rss+xml" href="/index.xml" />
           <link
             rel="preconnect"
@@ -289,67 +270,10 @@ export default class MyDocument extends Document {
               href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700&display=swap"
             />
           </noscript>
-          <link
-            rel="preload"
-            href="https://cdn.jsdelivr.net/npm/katex@0.13.0/dist/fonts/KaTeX_Main-Regular.woff2"
-            as="font"
-            type="font/woff2"
-            crossOrigin="anonymous"
-          />
-          <link
-            rel="preload"
-            href="https://cdn.jsdelivr.net/npm/katex@0.13.0/dist/fonts/KaTeX_Math-Italic.woff2"
-            as="font"
-            type="font/woff2"
-            crossOrigin="anonymous"
-          />
-          <link
-            rel="preload"
-            href="https://cdn.jsdelivr.net/npm/katex@0.13.0/dist/fonts/KaTeX_Size2-Regular.woff2"
-            as="font"
-            type="font/woff2"
-            crossOrigin="anonymous"
-          />
-          <link
-            rel="preload"
-            href="https://cdn.jsdelivr.net/npm/katex@0.13.0/dist/fonts/KaTeX_Size4-Regular.woff2"
-            type="font/woff2"
-            crossOrigin="anonymous"
-          />
-          <link
-            rel="stylesheet"
-            href="https://cdn.jsdelivr.net/npm/katex@0.13.0/dist/katex.min.css"
-            integrity="sha384-t5CR+zwDAROtph0PXGte6ia8heboACF9R5l/DiY+WZ3P2lxNgvJkQk5n7GPvLMYw"
-            crossOrigin="anonymous"
-          />
         </Head>
         <body className="antialiased text-black bg-white dark:bg-gray-900 dark:text-white">
           <Main />
           <NextScript />
-
-          <Script src="https://www.gstatic.com/firebasejs/8.1.1/firebase-app.js" />
-          <Script src="https://www.gstatic.com/firebasejs/8.1.1/firebase-analytics.js" />
-          {process.env.NODE_ENV === 'production' && (
-            <Script
-              dangerouslySetInnerHTML={{
-                __html: `
-            var firebaseConfig = {
-              apiKey: "AIzaSyDXDGGUots5JHk39kfGGV5ueRd09Ot3f50",
-              authDomain: "yceffort.firebaseapp.com",
-              databaseURL: "https://yceffort.firebaseio.com",
-              projectId: "yceffort",
-              storageBucket: "yceffort.appspot.com",
-              messagingSenderId: "754165146494",
-              appId: "1:754165146494:web:41d36183a76fb998f4892f",
-              measurementId: "G-PEKGCL9BKE"
-            };
-            // Initialize Firebase
-            firebase.initializeApp(firebaseConfig);
-            firebase.analytics();        
-          `,
-              }}
-            />
-          )}
         </body>
       </Html>
     )
